generation/requirements.txt
antlr4-python3-runtime==4.9.3
arrow==1.3.0
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colour==0.1.5
cpe==1.2.1
cybox==2.1.0.21
deepdiff==7.0.1
drawsvg==2.4.0
et-xmlfile==1.1.0
fqdn==1.5.1
idna==3.7
isoduration==20.11.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
loguru==0.7.2
lxml==5.2.2
maec==4.1.0.17
Markdown==3.6
markdown-it-py==3.0.0
mdurl==0.1.2
mitreattack-python==3.0.6
mixbox==1.0.5
netaddr==1.3.0
numpy==2.0.1
openpyxl==3.1.5
ordered-set==4.1.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
platformdirs==4.2.2
pluralizer==1.2.0
pooch==1.8.2
pycountry==24.6.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.1
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.19.1
setuptools==71.1.0
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
stix==1.2.0.11
stix2==3.0.1
stix2-elevator==4.1.7
stix2-patterns==2.0.0
stix2-validator==3.2.0
stixmarx==1.0.8
tabulate==0.9.0
taxii2-client==2.3.0
tqdm==4.66.4
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.2
weakrefmethod==1.0.3
webcolors==24.6.0
XlsxWriter==3.2.0
